feat(PL-396): remove api lib
