+++ apps/web-app/components/shared/directory/directory-sort/directory-sort.types.ts
@@ -6,3 +6,5 @@ export type TDirectorySortOption = typeof directorySortOptions[number];
 export interface IDirectorySortDropdownOption extends IDropdownOption {
   value: TDirectorySortOption;
 }
+
+export type TListSortDirection = 'asc' | 'desc';
