+++ apps/web-app/components/teams/teams-directory/teams-directory-list/teams-directory-list.tsx
@@ -1,4 +1,4 @@
-import { ITeam } from '@protocol-labs-network/api';
+import { ITeam } from '../../../../utils/teams.types';
 import { DirectoryList } from '../../../shared/directory/directory-list/directory-list';
 import { TeamCard } from '../team-card/team-card';
