--- libs/api/README.md
@@ -1,9 +1,0 @@
-# api
-
-Interfaces to be shared between frontend and backend apps for Protocol Labs Network.
-
-## Running unit tests
-
-Run `nx test api` to execute the unit tests via [Jest](https://jestjs.io).
-
-_This library was generated with [Nx](https://nx.dev)._
