--- libs/api/src/index.ts
@@ -1,3 +1,0 @@
-export * from './list';
-export * from './members';
-export * from './teams';
