--- libs/api/src/list.ts
@@ -1,13 +1,0 @@
-export type TListSortDirection = 'asc' | 'desc';
-
-export interface IListOptions {
-  sort?: IListSort[];
-  filterByFormula?: string;
-  fields?: string[];
-  pageSize?: number;
-}
-
-interface IListSort {
-  field: string;
-  direction: TListSortDirection;
-}
