+++ libs/members/data-access/src/index.spec.ts
@@ -1,5 +1,3 @@
-import { IMember } from '@protocol-labs-network/api';
-
 import { TMemberResponse } from '@protocol-labs-network/contracts';
 import { client } from '@protocol-labs-network/shared/data-access';
 import {
@@ -7,8 +5,6 @@ import {
   getMembers,
   getMembersFilters,
   getMemberUIDByAirtableId,
-  parseMember,
-  parseTeamMember,
 } from './index';
 
 jest.mock('@protocol-labs-network/shared/data-access', () => ({
@@ -135,260 +131,6 @@ describe('getMemberUIDByAirtableId', () => {
       query: { airtableRecId: airtableUID, select: 'uid' },
     });
     expect(response).toEqual(null);
-  });
-});
-
-describe('parseMember', () => {
-  it('should correctly parse a TMemberResponse into an IMember object', () => {
-    const memberResponse = {
-      ...memberResponseMock,
-      discordHandler: 'jsmith#1234',
-      githubHandler: 'jsmith',
-      image: { url: 'https://example.com/image.jpg' },
-      location: { country: 'USA', region: 'New York Region', city: 'New York' },
-      officeHours: 'https://example.com/office-hours',
-      skills: [
-        { uid: '', createdAt: '', updatedAt: '', title: 'JavaScript' },
-        { uid: '', createdAt: '', updatedAt: '', title: 'TypeScript' },
-      ],
-      teamMemberRoles: [
-        {
-          role: 'Developer',
-          team: { uid: 'team-1', name: 'Team 1' },
-          teamLead: true,
-          mainTeam: false,
-        },
-        {
-          role: 'Manager',
-          team: { uid: 'team-2', name: 'Team 2' },
-          teamLead: false,
-          mainTeam: true,
-        },
-      ],
-      twitterHandler: 'jsmith',
-    } as TMemberResponse;
-
-    const expectedResult: IMember = {
-      id: 'uid-john-smith',
-      name: 'John Smith',
-      email: 'john.smith@example.com',
-      image: 'https://example.com/image.jpg',
-      githubHandle: 'jsmith',
-      discordHandle: 'jsmith#1234',
-      twitter: 'jsmith',
-      officeHours: 'https://example.com/office-hours',
-      location: 'New York, USA',
-      skills: [
-        {
-          uid: '',
-          createdAt: '',
-          updatedAt: '',
-          title: 'JavaScript',
-        },
-        {
-          uid: '',
-          createdAt: '',
-          updatedAt: '',
-          title: 'TypeScript',
-        },
-      ],
-      teamLead: true,
-      teams: [
-        {
-          id: 'team-1',
-          name: 'Team 1',
-          role: 'Developer',
-          teamLead: true,
-          mainTeam: false,
-        },
-        {
-          id: 'team-2',
-          name: 'Team 2',
-          role: 'Manager',
-          teamLead: false,
-          mainTeam: true,
-        },
-      ],
-      mainTeam: {
-        id: 'team-2',
-        name: 'Team 2',
-        role: 'Manager',
-        teamLead: false,
-        mainTeam: true,
-      },
-    };
-
-    expect(parseMember(memberResponse)).toEqual(expectedResult);
-  });
-
-  it('should return correct defaults when optional fields are missing', () => {
-    const expectedOutput: IMember = {
-      id: 'uid-john-smith',
-      name: 'John Smith',
-      email: 'john.smith@example.com',
-      image: null,
-      githubHandle: null,
-      discordHandle: null,
-      twitter: null,
-      officeHours: null,
-      location: 'Not provided',
-      skills: [],
-      teamLead: false,
-      teams: [],
-      mainTeam: null,
-    };
-
-    expect(parseMember(memberResponseMock)).toEqual(expectedOutput);
-  });
-
-  it('should return correct location string when only country is provided', () => {
-    const memberResponse = {
-      ...memberResponseMock,
-      location: { country: 'USA' },
-    } as TMemberResponse;
-
-    const expectedOutput: IMember = {
-      id: 'uid-john-smith',
-      name: 'John Smith',
-      email: 'john.smith@example.com',
-      image: null,
-      githubHandle: null,
-      discordHandle: null,
-      twitter: null,
-      officeHours: null,
-      location: 'USA',
-      skills: [],
-      teamLead: false,
-      teams: [],
-      mainTeam: null,
-    };
-
-    expect(parseMember(memberResponse)).toEqual(expectedOutput);
-  });
-
-  it('should return correct location string when country and region are provided', () => {
-    const memberResponse = {
-      ...memberResponseMock,
-      location: { country: 'USA', region: 'New York Region' },
-    } as TMemberResponse;
-
-    const expectedOutput: IMember = {
-      id: 'uid-john-smith',
-      name: 'John Smith',
-      email: 'john.smith@example.com',
-      image: null,
-      githubHandle: null,
-      discordHandle: null,
-      twitter: null,
-      officeHours: null,
-      location: 'New York Region, USA',
-      skills: [],
-      teamLead: false,
-      teams: [],
-      mainTeam: null,
-    };
-
-    expect(parseMember(memberResponse)).toEqual(expectedOutput);
-  });
-});
-
-describe('parseTeamMember', () => {
-  const memberResponse = {
-    ...memberResponseMock,
-    discordHandler: 'jsmith#1234',
-    githubHandler: 'jsmith',
-    image: { url: 'https://example.com/image.jpg' },
-    location: { country: 'USA', region: 'New York Region', city: 'New York' },
-    officeHours: 'https://example.com/office-hours',
-    skills: [
-      { uid: '', createdAt: '', updatedAt: '', title: 'JavaScript' },
-      { uid: '', createdAt: '', updatedAt: '', title: 'TypeScript' },
-    ],
-    teamMemberRoles: [
-      {
-        role: 'Developer',
-        team: { uid: 'team-1', name: 'Team 1' },
-        teamLead: true,
-        mainTeam: false,
-      },
-      {
-        role: 'Manager',
-        team: { uid: 'team-2', name: 'Team 2' },
-        teamLead: false,
-        mainTeam: true,
-      },
-    ],
-    twitterHandler: 'jsmith',
-  } as TMemberResponse;
-
-  it('should correctly parse a TMemberResponse into an IMember object, excluding non-matching teams', () => {
-    const expectedResult: IMember = {
-      id: 'uid-john-smith',
-      name: 'John Smith',
-      email: 'john.smith@example.com',
-      image: 'https://example.com/image.jpg',
-      githubHandle: 'jsmith',
-      discordHandle: 'jsmith#1234',
-      twitter: 'jsmith',
-      officeHours: 'https://example.com/office-hours',
-      location: 'New York, USA',
-
-      skills: [
-        { uid: '', createdAt: '', updatedAt: '', title: 'JavaScript' },
-        { uid: '', createdAt: '', updatedAt: '', title: 'TypeScript' },
-      ],
-      teamLead: true,
-      teams: [
-        {
-          id: 'team-1',
-          name: 'Team 1',
-          role: 'Developer',
-          teamLead: true,
-          mainTeam: false,
-        },
-      ],
-      mainTeam: null,
-    };
-
-    expect(parseTeamMember(memberResponse, 'team-1')).toEqual(expectedResult);
-  });
-
-  it('should correctly parse a TMemberResponse into an IMember object, excluding non-matching teams', () => {
-    const expectedResult: IMember = {
-      id: 'uid-john-smith',
-      name: 'John Smith',
-      email: 'john.smith@example.com',
-      image: 'https://example.com/image.jpg',
-      githubHandle: 'jsmith',
-      discordHandle: 'jsmith#1234',
-      twitter: 'jsmith',
-      officeHours: 'https://example.com/office-hours',
-      location: 'New York, USA',
-
-      skills: [
-        { uid: '', createdAt: '', updatedAt: '', title: 'JavaScript' },
-        { uid: '', createdAt: '', updatedAt: '', title: 'TypeScript' },
-      ],
-      teamLead: false,
-      teams: [
-        {
-          id: 'team-2',
-          name: 'Team 2',
-          role: 'Manager',
-          teamLead: false,
-          mainTeam: true,
-        },
-      ],
-      mainTeam: {
-        id: 'team-2',
-        name: 'Team 2',
-        role: 'Manager',
-        teamLead: false,
-        mainTeam: true,
-      },
-    };
-
-    expect(parseTeamMember(memberResponse, 'team-2')).toEqual(expectedResult);
   });
 });
