+++ libs/teams/data-access/src/index.spec.ts
@@ -5,7 +5,6 @@ import {
   getTeams,
   getTeamsFilters,
   getTeamUIDByAirtableId,
-  parseTeam,
 } from './index';
 
 jest.mock('@protocol-labs-network/shared/data-access', () => ({
@@ -130,112 +129,6 @@ describe('getTeamUIDByAirtableId', () => {
       query: { airtableRecId: airtableUID, select: 'uid' },
     });
     expect(response).toEqual(null);
-  });
-});
-
-describe('parseTeam', () => {
-  it('should correctly parse a TTeamResponse object into an ITeam object', () => {
-    const teamResponse = {
-      ...teamResponseMock,
-      logo: { url: 'https://myteam.com/logo.png' },
-      website: 'https://myteam.com',
-      twitterHandler: '@myteam',
-      shortDescription: 'We build cool stuff',
-      longDescription:
-        'We are a team of skilled engineers who build innovative products',
-      technologies: [{ title: 'Filecoin' }, { title: 'IPFS' }],
-      membershipSources: [
-        { title: 'Membership Source A' },
-        { title: 'Membership Source B' },
-      ],
-      industryTags: [
-        {
-          uid: '',
-          createdAt: '',
-          updatedAt: '',
-          title: 'Software Development',
-          industryCategoryUid: '',
-        },
-        {
-          uid: '',
-          createdAt: '',
-          updatedAt: '',
-          title: 'Blockchain',
-          industryCategoryUid: '',
-        },
-      ],
-      fundingStage: { title: 'Seed' },
-      teamMemberRoles: [{ member: { uid: '456' } }, { member: { uid: '789' } }],
-      contactMethod: 'https://myteam.com/contact',
-    } as TTeamResponse;
-
-    const expectedResult = {
-      id: 'team-01',
-      name: 'Team 01',
-      logo: 'https://myteam.com/logo.png',
-      website: 'https://myteam.com',
-      twitter: '@myteam',
-      shortDescription: 'We build cool stuff',
-      longDescription:
-        'We are a team of skilled engineers who build innovative products',
-      technologies: [
-        {
-          title: 'Filecoin',
-        },
-        {
-          title: 'IPFS',
-        },
-      ],
-      fundingStage: 'Seed',
-      industryTags: [
-        {
-          createdAt: '',
-          industryCategoryUid: '',
-          title: 'Software Development',
-          uid: '',
-          updatedAt: '',
-        },
-        {
-          createdAt: '',
-          industryCategoryUid: '',
-          title: 'Blockchain',
-          uid: '',
-          updatedAt: '',
-        },
-      ],
-      membershipSources: [
-        {
-          title: 'Membership Source A',
-        },
-        {
-          title: 'Membership Source B',
-        },
-      ],
-      members: ['456', '789'],
-      contactMethod: 'https://myteam.com/contact',
-    };
-
-    expect(parseTeam(teamResponse)).toEqual(expectedResult);
-  });
-
-  it('should correctly parse a TTeamResponse object into an ITeam object', () => {
-    const expectedResult = {
-      id: 'team-01',
-      name: 'Team 01',
-      logo: null,
-      website: null,
-      twitter: null,
-      shortDescription: null,
-      longDescription: null,
-      technologies: [],
-      fundingStage: null,
-      membershipSources: [],
-      industryTags: [],
-      members: [],
-      contactMethod: null,
-    };
-
-    expect(parseTeam(teamResponseMock)).toEqual(expectedResult);
   });
 });
