+++ libs/teams/data-access/src/index.ts
@@ -1,4 +1,3 @@
-import { ITeam } from '@protocol-labs-network/api';
 import { TTeamResponse } from '@protocol-labs-network/contracts';
 import {
   client,
@@ -37,53 +36,6 @@ export const getTeamUIDByAirtableId = async (id: string) => {
   });
 
   return res.status === 200 && res.body[0] ? res.body[0].uid : null;
-};
-
-/**
- * Parse team fields values into a team object.
- **/
-export const parseTeam = (team: TTeamResponse): ITeam => {
-  const {
-    uid: id,
-    name,
-    logo,
-    website,
-    twitterHandler: twitter,
-    shortDescription,
-    longDescription,
-    technologies,
-    membershipSources,
-    industryTags,
-    fundingStage,
-    teamMemberRoles,
-    contactMethod,
-  } = team;
-
-  const memberIds = teamMemberRoles?.length
-    ? [
-        ...new Set(
-          teamMemberRoles.map(
-            (teamMemberRole) => teamMemberRole.member?.uid || ''
-          )
-        ),
-      ]
-    : [];
-
-  return {
-    id,
-    name,
-    logo: logo?.url || null,
-    website: website || null,
-    twitter: twitter || null,
-    shortDescription: shortDescription || null,
-    longDescription: longDescription || null,
-    technologies: technologies || [],
-    fundingStage: fundingStage?.title || null,
-    industryTags: industryTags || [],
-    membershipSources: membershipSources || [],
-    members: memberIds,
-    contactMethod: contactMethod || null,
-  };
 };
 
 /**
