+++ tsconfig.base.json
@@ -16,7 +16,6 @@
     "baseUrl": ".",
     "paths": {
       "@protocol-labs-network/airtable": ["libs/airtable/src/index.ts"],
-      "@protocol-labs-network/api": ["libs/api/src/index.ts"],
       "@protocol-labs-network/contracts": ["libs/contracts/src/index.ts"],
       "@protocol-labs-network/funding-stages/data-access": [
         "libs/funding-stages/data-access/src/index.ts"
